Fixes bug in login API call methods from (was using `data` rather than `json` format).
This commit updates the version of the Umami package to 0.2.18 and modifies the login
functions in the `umami/impl/__init__.py` file to use JSON instead of form-encoded data for API
requests. This change is part of a larger effort to improve the reliability and security of the
package's communication with its backend services.